requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
streamlit
langchain_core
